third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-evaluate-without-enabling.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

window.dummyObject = { a : 1 };

function test()
{
    InspectorTest.sendCommand("Runtime.evaluate", { "expression": "window.dummyObject" }, evaluateCallback);

    function evaluateCallback(result)
    {
        InspectorTest.navigate("runtime-evaluate-without-enabling.html", reloadCallback.bind(null, result.result.result.objectId));
    }

    function reloadCallback(remoteObjectId)
    {
        InspectorTest.sendCommand("Runtime.getProperties", { objectId: remoteObjectId, ownProperties: true }, getPropertiesCallback);
    }

    function getPropertiesCallback(result)
    {
        result.id = 0;
        InspectorTest.logObject(result);
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onLoad="runTest();">
Tests that default execution context accessed without enabling Runtime domain gets properly cleaned up on reload.
</body>
</html>
