third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-evaluate-without-enabling-expected.txt

Tests that default execution context accessed without enabling Runtime domain gets properly cleaned up on reload.
Page reloaded.
{
    error : {
        code : -32000
        message : Inspected frame has gone
    }
    id : 0
}
